renaming files: checkpoint
