+++ components/country/AppsStatsCircumventionRow.js
