+++ pages/about.js
@@ -8,7 +8,7 @@ import {
 } from 'ooni-components'
 
 import Layout from '../components/Layout'
-import NavBar from '../components/nav-bar'
+import NavBar from '../components/NavBar'
 
 export default class About extends React.Component {
   render () {
